Added runtime check using Tcl_GetCurrentThread() to autodetect a threaded Tcl library when compiled and linked against Tcl 8.1 and later.
Replaced EGG_TCL_ENABLE_THREADS and EGG_TCL_CHECK_THREADS autoconf macros with EGG_TCL_CHECK_GETCURRENTTHREAD.

+++ src/bg.c
@@ -3,7 +3,7 @@
  *   moving the process to the background, i.e. forking, while keeping threads
  *   happy.
  *
- * $Id: bg.c,v 1.15 2010/01/03 13:27:31 pseudo Exp $
+ * $Id: bg.c,v 1.16 2010/01/26 03:12:15 tothwolf Exp $
  */
 /*
  * Copyright (C) 1997 Robey Pointer
@@ -28,13 +28,7 @@
 #include <signal.h>
 #include "bg.h"
 
-#ifdef HAVE_TCL_THREADS
-#  define SUPPORT_THREADS
-#endif
-
 extern char pid_file[];
-
-#ifdef SUPPORT_THREADS
 
 /* When threads are started during eggdrop's init phase, we can't simply
  * fork() later on, because that only copies the VM space over and
@@ -100,8 +94,6 @@ typedef struct {
 
 static bg_t bg = { 0 };
 
-#endif /* SUPPORT_THREADS */
-
 
 /* Do everything we normally do after we have split off a new
  * process to the background. This includes writing a PID file
@@ -136,7 +128,9 @@ static void bg_do_detach(pid_t p)
 
 void bg_prepare_split(void)
 {
-#ifdef SUPPORT_THREADS
+  if (!tcl_threaded())
+    return;
+
   /* Create a pipe between parent and split process, fork to create a
    * parent and a split process and wait for messages on the pipe. */
   pid_t p;
@@ -188,10 +182,8 @@ error:
   /* We only reach this point in case of an error.
    */
   fatal("COMMUNICATION THROUGH PIPE BROKE.", 0);
-#endif
 }
 
-#ifdef SUPPORT_THREADS
 /* Transfer contents of pid_file to parent process. This is necessary,
  * as the pid_file[] buffer has changed in this fork by now, but the
  * parent needs an up-to-date version.
@@ -213,11 +205,12 @@ static void bg_send_pidfile(void)
 error:
   fatal("COMMUNICATION THROUGH PIPE BROKE.", 0);
 }
-#endif
 
 void bg_send_quit(bg_quit_t q)
 {
-#ifdef SUPPORT_THREADS
+  if (!tcl_threaded())
+    return;
+
   if (bg.state == BG_PARENT) {
     kill(bg.child_pid, SIGKILL);
   } else if (bg.state == BG_SPLIT) {
@@ -232,21 +225,21 @@ void bg_send_quit(bg_quit_t q)
     if (write(bg.comm_send, &message, sizeof(message)) < 0)
       fatal("COMMUNICATION THROUGH PIPE BROKE.", 0);
   }
-#endif
 }
 
 void bg_do_split(void)
 {
-#ifdef SUPPORT_THREADS
-  /* Tell our parent process to go away now, as we don't need it anymore. */
-  bg_send_quit(BG_QUIT);
-#else
-  /* Split off a new process. */
-  int xx = fork();
+  if (tcl_threaded()) {
+    /* Tell our parent process to go away now, as we don't need it anymore. */
+    bg_send_quit(BG_QUIT);
 
-  if (xx == -1)
-    fatal("CANNOT FORK PROCESS.", 0);
-  if (xx != 0)
-    bg_do_detach(xx);
-#endif
+  } else {
+    /* Split off a new process. */
+    int xx = fork();
+
+    if (xx == -1)
+      fatal("CANNOT FORK PROCESS.", 0);
+    if (xx != 0)
+      bg_do_detach(xx);
+  }
 }
